Refactor env into a separate file
