+++ src/ops/store.rs
@@ -7,7 +7,7 @@
 use crate::call_context::CallContext;
 use crate::VMError;
 
-use crate::resolver::Env;
+use crate::env::Env;
 use canonical::{Canon, IdHash, Sink, Source, Store};
 
 pub struct Get;
